+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed all upload commands to upload as binary instead of XML.
 * Added support for CSRF negotiation to fix asset uploading. ([#25][#25])
 * Added support for Vector3int16.
+* Added support for BinaryString values.
 
 [#25]: https://github.com/rojo-rbx/remodel/issues/25